refactor: clean files
