+++ src/components/main/Canvas.tsx
@@ -13,7 +13,7 @@ import {
   DEFAULT_BACKGROUND_ENABLED,
   DEFAULT_PERMISSION,
 } from '../../config/settings';
-import { APP_SETTINGS } from '../../constants/constants';
+import { APP_SETTINGS } from '../../config/constants';
 import Settings from '../modes/teacher/Settings';
 import BackgroundImage from './BackgroundImage';
 import CanvasScaleControl from './CanvasScaleControl';
