+++ src/components/modes/teacher/ImageUpload.tsx
@@ -14,7 +14,7 @@ import { FileInput } from '@uppy/react';
 
 import { MUTATION_KEYS, hooks, useMutation } from '../../../config/queryClient';
 import { MAX_FILE_SIZE } from '../../../config/settings';
-import { APP_SETTINGS } from '../../../constants/constants';
+import { APP_SETTINGS } from '../../../config/constants';
 import '../../../index.css';
 import configureUppy from '../../../utils/uppy';
 import { useAppSettingContext } from '../../context/AppSettingContext';
