+++ src/config/selectors.ts
@@ -1,4 +1,5 @@
-export const dataCyWrapper = (selector) => `[data-cy="${selector}"]`;
+export const dataCyWrapper = (selector: string): string =>
+  `[data-cy="${selector}"]`;
 export const SETTINGS_BUTTON_CY = 'settingsButton';
 export const SETTINGS_CY = 'settings';
 export const SCROLL_CONTAINER_CY = 'scrollContainer';
