--- src/constants/langs.js
@@ -1,6 +1,0 @@
-// todo remove suppression on multiple exports
-// eslint-disable-next-line import/prefer-default-export
-export const options = [
-  { value: 'en', label: 'English' },
-  { value: 'fr', label: 'French' },
-];
